twitter_simulation/requirements.txt
google-cloud-pubsub
google-cloud-secret-manager
requests